fastapi-optimized-project/pyproject.toml
[tool.poetry]
name = "fastapi-optimized"
version = "0.1.0"
description = "A highly optimized FastAPI project with best practices"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
redis = "^5.0.1"
httpx = "^0.27.2"
python-memcached = "^1.61"
pydantic-settings = "^2.10.1"
feedparser = "^6.0.11"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
apscheduler = "^3.10.4"
croniter = "^1.4.1"
alembic = "^1.13.1"
aiosqlite = "^0.20.0"
crawl4ai = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.27.2"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
